fix: changed id to int

+++ CarPooling/app/src/main/java/it/polito/mad/group27/carpooling/ui/trip/Trip.kt
@@ -67,7 +67,7 @@ object BigDecimalSerializer: KSerializer<BigDecimal> {
 @Serializable
 @Parcelize
 data class Trip(
-    var id: Long = -1,
+    var id: Int = -1,
     @Serializable(with=UriSerializer::class)
     var carImageUri: Uri? = null,
     @Serializable(with=DateSerializer::class)
